Rename test-ovs-icmp-frag-offload.sh to test-ovs-icmp-frag.sh
Also disable testing for offload as we currently don't support
offloading of frag first/later.

+++ test-ovs-icmp-frag.sh
@@ -9,6 +9,10 @@
 my_dir="$(dirname "$0")"
 . $my_dir/common.sh
 
+# we currently don't support offloading of frag first/later
+# so testing for offload is disabled.
+TEST_OFFLOAD=0
+
 test -z "$VF" && fail "Missing VF"
 test -z "$VF2" && fail "Missing VF2"
 test -z "$REP" && fail "Missing REP"
@@ -66,10 +70,12 @@ start_check_syndrome
 # quick ping to make ovs add rules
 ping -q -c 1 -w 1 $VM2_IP -s 2000 && success || err
 
-tdtmpfile=/tmp/$$.pcap
-timeout 15 tcpdump -nnepi $REP icmp -c 30 -w $tdtmpfile &
-tdpid=$!
-sleep 0.5
+if [ "$TEST_OFFLOAD" = 1 ]; then
+    tdtmpfile=/tmp/$$.pcap
+    timeout 15 tcpdump -nnepi $REP icmp -c 30 -w $tdtmpfile &
+    tdpid=$!
+    sleep 0.5
+fi
 
 title "Test ping $VM1_IP -> $VM2_IP - expect to pass"
 ping -q -c 30 -i 0.2 -w 15 -s 2000 $VM2_IP && success || err
@@ -80,15 +86,17 @@ check_offloaded_rules 4
 ovs-dpctl dump-flows type=offloaded --names
 tc -s filter show dev $REP ingress
 
-kill $tdpid 2>/dev/null
-sleep 1
-count=`tcpdump -nr $tdtmpfile | wc -l`
-title "Verify with tcpdump"
-if [[ $count -gt 2 ]]; then
-    err "No offload"
-    tcpdump -nr $tdtmpfile
-else
-    success
+if [ "$TEST_OFFLOAD" = 1 ]; then
+    kill $tdpid 2>/dev/null
+    sleep 1
+    count=`tcpdump -nr $tdtmpfile | wc -l`
+    title "Verify with tcpdump"
+    if [[ $count -gt 2 ]]; then
+        err "No offload"
+        tcpdump -nr $tdtmpfile
+    else
+        success
+    fi
 fi
 
 check_syndrome || err
